added background image deleted screens
